+++ R/preamble.R
@@ -17,12 +17,6 @@ if(file.exists(the$constants_fn)) {
   # What is not in constatns
   the$assetlist <- data.table(listnm=c(rep("defaultIdx",2),rep("ix2",2)),ticker=c("SPY","QQQ","EWZ","EEM"))
   the$cachedir <- the$defaultcachedir
-  u1 <- avsd$defaults[varType=="cache",]
-  u1 <- lapply(seq(1,nrow(u1)), \(i) assign(u1[i,]$var, paste0(the$cachedir,"/",u1[i,]$value_str), envir=the))
-  u1 <- avsd$defaults[varType=="str",]
-  u1 <- lapply(seq(1,nrow(u1)), \(i) assign(u1[i,]$var, u1[i,]$value_str, envir=the))
-  u1 <- avsd$defaults[varType=="log",]
-  u1 <- lapply(seq(1,nrow(u1)), \(i) assign(u1[i,]$var, u1[i,]$value_log, envir=the))
 }
 
 .datatable.aware = TRUE